Fix STD instruction overwriting register object with numeric value
Without knowing much about the overall structure, it appears that DMEM is a bunch of Register objects. There are certain areas where the objects are overwritten with something like:
`this.getDMEM()[k] = Rd`
This overwrites the Register object and the app crashes in `populateRegisters()` because it expects the `.getValue()` function to exist for the members of the array being iterated on.

There are other areas where DMEM is written to directly instead of using `.setValue()` but I did not want to meddle with parts of the code I don't know how to test.