Fix bad `current_tok` references
In a couple of error messages, code inside template literals was referencing the `current_tok` variable which did not exist, which led to uncaught exceptions in the simulator. This meant it stayed stuck on whatever previous state it was on, with no user-visible error messages.

Fixed by updating the template to use `this.current_tok` instead.